PlantMedic-Web/crop_disease_model_app/requirements.txt
Flask==1.1.2
numpy==1.17.3
Pillow==6.2.1
tensorflow==2.4.1
Werkzeug==1.0.1

